# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.BulkMarkAttendanceRequest:
    properties:
      user_ids:
        items:
          type: string
        type: array
    required:
    - user_ids
    type: object
...
  request.MarkAttendanceRequest:
    properties:
      notes:
        type: string
      user_id:
        type: string
    required:
    - user_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@eventcampus.com
    name: API Support
  description: |-
    REST API for Event Campus - Campus Event Management System

    Features:
    - User authentication (Register, Login)
    - Event management (Create, Read, Update, Delete)
    - Event registration
    - Attendance tracking
    - Whitelist/Organizer approval system
    - Email notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Event Campus API
  version: "1.0"
paths:
...
  /events/{id}/attendance:
    get:
      consumes:
      - application/json
      description: Get attendance list for a specific event (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attendance retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID or failed to get attendance
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get event attendance
      tags:
      - Attendance
    post:
      consumes:
      - application/json
      description: Mark attendance for a single user at an event (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Attendance details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.MarkAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Attendance marked successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or failed to mark
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark single attendance
      tags:
      - Attendance
  /events/{id}/attendance/bulk:
    post:
      consumes:
      - application/json
      description: Mark attendance for multiple users at once (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: List of user IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.BulkMarkAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Bulk attendance marked successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or failed to mark
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark bulk attendance
      tags:
      - Attendance
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"